Increase support >:|

+++ miniaudio-sys/src/constants.rs
@@ -17,14 +17,14 @@ pub const MA_SAMPLE_RATE_352800: c_int = 352800;
 pub const MA_SAMPLE_RATE_384000: c_int = 384000;
 
 /// For simplicity, miniaudio does not support PCM samples that are not byte aligned.
-pub const MA_MIN_PCM_SAMPLE_SIZE_IN_BYTES: c_int = 1;
-pub const MA_MAX_PCM_SAMPLE_SIZE_IN_BYTES: c_int = 8;
-pub const MA_MIN_CHANNELS: c_int = 1;
-pub const MA_MAX_CHANNELS: c_int = 32;
+pub const MA_MIN_PCM_SAMPLE_SIZE_IN_BYTES: usize = 1;
+pub const MA_MAX_PCM_SAMPLE_SIZE_IN_BYTES: usize = 8;
+pub const MA_MIN_CHANNELS: usize = 1;
+pub const MA_MAX_CHANNELS: usize = 32;
 pub const MA_MIN_SAMPLE_RATE: c_int = MA_SAMPLE_RATE_8000;
 pub const MA_MAX_SAMPLE_RATE: c_int = MA_SAMPLE_RATE_384000;
-pub const MA_SRC_SINC_MIN_WINDOW_WIDTH: c_int = 2;
-pub const MA_SRC_SINC_MAX_WINDOW_WIDTH: c_int = 32;
-pub const MA_SRC_SINC_DEFAULT_WINDOW_WIDTH: c_int = 32;
-pub const MA_SRC_SINC_LOOKUP_TABLE_RESOLUTION: c_int = 8;
-pub const MA_SRC_INPUT_BUFFER_SIZE_IN_SAMPLES: c_int = 256;
+pub const MA_SRC_SINC_MIN_WINDOW_WIDTH: usize = 2;
+pub const MA_SRC_SINC_MAX_WINDOW_WIDTH: usize = 32;
+pub const MA_SRC_SINC_DEFAULT_WINDOW_WIDTH: usize = 32;
+pub const MA_SRC_SINC_LOOKUP_TABLE_RESOLUTION: usize = 8;
+pub const MA_SRC_INPUT_BUFFER_SIZE_IN_SAMPLES: usize = 256;
